Split profile/module/extensions
